Consistant formatting, single file, easier WATER_TIME config
There were a few formatting inconsistencies like indention or using
grave instead $() for command substitution.

All information is now kept in a single file $HOME/.water. The first
line is the true/false, the second is the unix timestamp.

Replaced all conditionals with POSIX style statements. This should make
it compatible with more POSIX shells.

The WATER_TIME variable will take an already defined value and fallback
to the default of 10800. This allows a user to set the variable outside
of the script and not have to mess with it at all.

The .water file will be made automatically if it doesn't exist. The
first time I ran it, I got an err from cat because the file didn't
exist yet. That will no longer be the case.