Better tests, another file reduction, added next_drink
The tests actually check for proper output and not just display it.

The true/false piece has been removed. Now the last drink time
is saved so the next drink time can be calculated dynamically.

Added next_drink function to display the time of the next drink.

The time file is now configurable by setting the variable
DRINK_WATER_CONF before calling in the script. This was mainly for
testing, but can be useful.